--- a/spec/features/search_spec.rb
+++ b/spec/features/search_spec.rb
@@ -29,8 +29,18 @@ RSpec.feature 'searching' do
       end
     end
 
-    it "only searches all" do
+    it "only searches all and does not display search options for dashboard files" do
       visit '/'
+
+      # it "does not display search options for dashboard files" do
+      # This section was tested on its own, and required a full setup.
+      within(".input-group-btn") do
+        expect(page).not_to have_content("My Works")
+        expect(page).not_to have_content("My Collections")
+        expect(page).not_to have_content("My Shares")
+      end
+      # end
+
       expect(page).to have_content("All")
       expect(page).to have_css("a[data-search-label*=All]", visible: false)
       expect(page).not_to have_css("a[data-search-label*='My Works']", visible: false)
@@ -50,15 +60,6 @@ RSpec.feature 'searching' do
 
       expect(page.body).to include "<span itemprop=\"keywords\"><a href=\"/catalog?f%5Bkeyword_sim%5D%5B%5D=taco&amp;locale=en\">taco</a></span>"
       expect(page.body).to include "<span itemprop=\"keywords\"><a href=\"/catalog?f%5Bkeyword_sim%5D%5B%5D=mustache&amp;locale=en\">mustache</a></span>"
-    end
-
-    it "does not display search options for dashboard files" do
-      visit "/"
-      within(".input-group-btn") do
-        expect(page).not_to have_content("My Works")
-        expect(page).not_to have_content("My Collections")
-        expect(page).not_to have_content("My Shares")
-      end
     end
   end
 end