## Consolidating related features into single `it`

f9457393aa9f9907aa1d45a498fef6ae828b9d82

Each `it` block comes with a setup time. Feature spec setup times are
even more expensive.

Before the change: `Finished in 40.03 seconds`
After the change: `Finished in 28.14 seconds`

Feature specs have a higher variance on completion time (as there is
browser interaction).

## Consolidating related features into single `it`

db57aa80399360b4bba3130b7842beb9a05a6ba6

Each `it` block comes with a setup time. Feature spec setup times are
even more expensive.

Before the change: `Finished in 6.17 seconds`
After the change: `Finished in 5.33 seconds`

It's a small performance gain, but each second matters.

Feature specs have a higher variance on completion time (as there is
browser interaction).